backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain==0.3.0
langchain-community
langchain-ollama
chromadb
pypdf
python-docx
python-dotenv
langchain-text-splitters
